test: move TokenStream display tests to separate file
